requirements.txt
Pillow == 4.0.0
python-telegram-bot == 5.3.0
